Worked with Anton Dahlin on the following code:
Manually start intake and automatically stop intake when note is detected. Then manually start the process to transfer the note from intake to shooter section and automatically stop when note is detected. Finally, eject the note which would happen to score in the amp.

Deleted the "simulink/slprj" folder.

Swapped the left/right shooter CAN IDs and swapped the negative signs in the Stateflow code.