Tuning update for the intake to shooter transfer.
This is as good as we can get with a software only solution.
Intake at 45 degrees
Shooter at 35 degrees
9.75 inch gap from center of intake roller to edge of shooter wheels
Shooter height adjusted for an easy transition without hitting the polycarb ramp.

For better performance we need a mechanical stop.  We are looking into putting a servo on the shooter that can rotate an arm into position to stop the note from pushing out too far.