Tuning updates from match practice at AC.
Disconnected April tag trap alignment joystick button input since it was not working.
Moved buttons for April tag speaker and amp alignment.
Added an  override for test mode to move the servo to a down position.
Attempted to fix issue with swerve snapping back to 0 degrees after using speaker or amp alignment, but it is not working yet.